+++ pkg/config/config.go
@@ -162,6 +162,7 @@ type Configuration interface {
 
 // configuration stores the configuration
 type configuration struct {
+	skipResourceFilters           bool
 	defaultRegistry               string
 	enableDefaultRegistryMutation bool
 	excludedGroups                []string
@@ -177,8 +178,9 @@ type configuration struct {
 }
 
 // NewDefaultConfiguration ...
-func NewDefaultConfiguration() *configuration {
+func NewDefaultConfiguration(skipResourceFilters bool) *configuration {
 	return &configuration{
+		skipResourceFilters:           skipResourceFilters,
 		defaultRegistry:               "docker.io",
 		enableDefaultRegistryMutation: true,
 		excludedGroups:                defaultExcludedGroups,
@@ -187,8 +189,8 @@ func NewDefaultConfiguration() *configuration {
 }
 
 // NewConfiguration ...
-func NewConfiguration(client kubernetes.Interface) (Configuration, error) {
-	cd := NewDefaultConfiguration()
+func NewConfiguration(client kubernetes.Interface, skipResourceFilters bool) (Configuration, error) {
+	cd := NewDefaultConfiguration(skipResourceFilters)
 	if cm, err := client.CoreV1().ConfigMaps(kyvernoNamespace).Get(context.TODO(), kyvernoConfigMapName, metav1.GetOptions{}); err != nil {
 		if !errors.IsNotFound(err) {
 			return nil, err
@@ -202,6 +204,7 @@ func NewConfiguration(client kubernetes.Interface) (Configuration, error) {
 func (cd *configuration) ToFilter(gvk schema.GroupVersionKind, subresource, namespace, name string) bool {
 	cd.mux.RLock()
 	defer cd.mux.RUnlock()
+	if !cd.skipResourceFilters {
 	for _, f := range cd.filters {
 		if wildcard.Match(f.Group, gvk.Group) &&
 			wildcard.Match(f.Version, gvk.Version) &&
@@ -218,6 +221,12 @@ func (cd *configuration) ToFilter(gvk schema.GroupVersionKind, subresource, name
 				wildcard.Match(f.Kind, gvk.Kind) &&
 				wildcard.Match(f.Namespace, name) {
 				return true
+			}
+			if kind == "Namespace" {
+				// [Namespace,kube-system,*] || [*,kube-system,*]
+				if (f.Kind == "Namespace" || f.Kind == "*") && wildcard.Match(f.Namespace, name) {
+					return true
+				}
 			}
 		}
 	}